requirements.txt
asgiref==3.11.0
Django==4.2.26
django-filter==25.1
django-phonenumber-field==8.3.0
django-resized==1.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
phonenumbers==9.0.19
pillow==11.3.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
